requirements.txt
certifi==2020.12.5
facebook-sdk==3.1.0
PySocks==1.7.1
requests-oauthlib==1.3.0
tweepy==3.10.0
